+++ CHANGELOG.md
@@ -22,7 +22,9 @@ Now, instead, the "Executor" classes implement the `getBinary()` method, which r
 * the `AbstractBackupUtility::_getDriver()` had been deprecated and has now been removed;
 * the `AbstractExecutor::__call()` magic method, which provided backwards compatibility for deprecated
   `getExportExecutable()` and `getImportExecutable()` methods, has been removed;
-* the `AbstractBackupUtility::__get()` magic method, which allowed to read the properties of the "Executor" classes, had
+* the `beforeExport()` and `beforeImport()` event methods provided by `AbstractExecutor`, `MysqlExecutor` and
+  `SqliteExecutor` now explicitly return `void`;
+* the `AbstractBackupUtility::__get()` magic method, which allowed reading the properties of the "Executor" classes, had
   been deprecated and has now been removed;
 * the `DatabaseBackup.connection` configuration name had been deprecated and now have been removed;
 * configuration names `DatabaseBackup.mysql`, `DatabaseBackup.postgres` and `DatabaseBackup.sqlite` had been deprecated,